+++ scurrypy/core/gateway.py
@@ -6,6 +6,8 @@ import logging
 
 logger = logging.getLogger(__name__)
 
+MIN_BACKOFF = 5
+
 class GatewayClient:
     def __init__(self, gateway_url: str, shard_id: int, total_shards: int):
         """Initialize this websocket.
@@ -21,6 +23,7 @@ class GatewayClient:
         self.seq = None
         self.session_id = None
         self.allow_resume = False # default: assume IDENTIFY
+        self.backoff = MIN_BACKOFF
         self.heartbeat_task = None
         self.heartbeat_interval = None
         self.event_queue = asyncio.Queue()
@@ -28,31 +31,33 @@ class GatewayClient:
         self.base_url = gateway_url
         self.url_params = "?v=10&encoding=json"
 
-    async def wait_reconnect(self, backoff: float):
-        import random
+    async def wait_reconnect(self):
+        """Sleep for exponentially increasing time between reconnects."""
 
-        backoff = min(backoff * 2, 60)
-        backoff *= random.uniform(0.8, 1.2)
+        self.backoff = min(self.backoff * 2, 60)
         
-        logger.warning(f"SHARD ID {self.shard_id}: Disconnected, reconnecting in {backoff}s...")
-        await asyncio.sleep(backoff)
+        logger.warning(f"SHARD ID {self.shard_id}: Disconnected, reconnecting in {self.backoff}s...")
+        await asyncio.sleep(self.backoff)
 
-        return backoff
+    async def start(self, token: str, intents: int):
+        """Start this websocket's connection.
 
-    async def start(self, token, intents):
-        backoff = 5
-
+        Args:
+            token (str): the bot's token
+            intents (int): the bot's intents
+        """
         while True:
             try:
                 await self.connect_ws()
 
-                if self.allow_resume and self.session_id and self.seq:
+                if self.allow_resume and self.session_id:
+                    logger.debug(f"SHARD ID {self.shard_id}: Attempting to resume...")
                     await self.resume(token)
                 else:
+                    logger.debug(f"SHARD ID {self.shard_id}: Attempting to identify...")
                     await self.identify(token, intents)
 
                 await self._listen()  # blocks until disconnect
-                backoff = 5  # reset after successful connection
 
             except websockets.exceptions.ConnectionClosedOK:
                 logger.info(f"SHARD ID {self.shard_id}: Connection closed properly.")
@@ -62,13 +67,13 @@ class GatewayClient:
                 logger.error(f"SHARD ID {self.shard_id}: {e}")
                 await self.close_ws()
 
-                backoff = await self.wait_reconnect(backoff)
+                await self.wait_reconnect()
 
             except Exception:
                 logger.exception(f"SHARD ID {self.shard_id}: Unexpected error")
                 await self.close_ws()
                 
-                backoff = await self.wait_reconnect(backoff)
+                await self.wait_reconnect()
 
     async def connect_ws(self):
         """Connect to Discord's Gateway (websocket)."""
@@ -88,11 +93,19 @@ class GatewayClient:
         self.heartbeat_task = asyncio.create_task(self.heartbeat())
 
     async def send(self, data: dict):
-        """Send data through the websocket."""
+        """Send data through the websocket.
+
+        Args:
+            data (dict): data to send
+        """
         await self.ws.send(json.dumps(data))
 
     async def receive(self):
-        """Receive data through the websocket."""
+        """Receive data through the websocket.
+
+        Returns:
+            (dict): websocket data
+        """
         return json.loads(await self.ws.recv())
 
     async def heartbeat(self):
@@ -109,7 +122,15 @@ class GatewayClient:
             await asyncio.sleep(self.heartbeat_interval)
 
     async def identify(self, token: str, intents: int):
-        """Send an IDENTIFY payload to handshake for bot."""
+        """Send an IDENTIFY payload to handshake for bot.
+
+        Args:
+            token (str): the bot's token
+            intents (int): the bot's intents
+
+        Raises:
+            (ConnectionError): invalid intents
+        """
 
         if not isinstance(intents, int):
             raise ConnectionError("Invalid intents.")
@@ -130,20 +151,28 @@ class GatewayClient:
         logger.info(f"SHARD ID {self.shard_id}: IDENIFY Sent.")
 
     async def resume(self, token: str):
-        """Send a RESUME payload to resume a connection."""
+        """Send a RESUME payload to resume a connection.
+
+        Args:
+            token (str): the bot's token
+        """
 
         await self.send({
             'op': 6,
             'd': {
                 'token': f"Bot {token}",
                 'session_id': self.session_id,
-                'seq': self.seq
+                'seq': self.seq or 0
             }
         })
         logger.info(f"SHARD ID {self.shard_id}: Resume Sent.")
 
     async def _listen(self):
-        """Listen for events and queue them to be picked up by Client."""
+        """Listen for events and queue them to be consumed by Client.
+
+        Raises:
+            (ConnectionError): an error occurred
+        """
 
         while self.ws:
             data = await self.receive()
@@ -152,24 +181,34 @@ class GatewayClient:
             match op_code:
                 case 0:  # DISPATCH
                     self.seq = data.get("s") or self.seq
+                    event_data = data.get('d')
                     dispatcher_type = data.get("t")
 
                     if dispatcher_type == "READY":
-                        self.session_id = data["d"].get("session_id")
-                        self.base_url = data["d"].get("resume_gateway_url", self.base_url)
+                        self.session_id = event_data.get("session_id")
+                        self.base_url = event_data.get("resume_gateway_url", self.base_url)
 
-                    await self.event_queue.put((dispatcher_type, data["d"]))
+                        # this is a stable connection so reset backoff
+                        self.backoff = MIN_BACKOFF
+
+                    await self.event_queue.put((dispatcher_type, event_data))
 
                 case 7:  # RECONNECT
                     self.allow_resume = True
-                    raise ConnectionError("Reconnect requested by server")
+                    logger.debug(f"SHARD ID {self.shard_id}: Reconnect requested by server.")
+                    raise ConnectionError("Reconnect requested by server.")
 
-                case 9:  # INVALID_SESSION
-                    resumable = data.get("d", False)
+                case 9:
+                    resumable = bool(data.get("d"))
                     self.allow_resume = resumable
-                    if not resumable:
+
+                    if resumable:
+                        logger.debug(f"SHARD ID {self.shard_id}: Invalid session (resumable).")
+                    else:
                         self.session_id = self.seq = None
-                    raise ConnectionError("Invalid session")
+                        logger.debug(f"SHARD ID {self.shard_id}: Invalid session (not resumable).")
+
+                    raise ConnectionError("Invalid session.")
 
                 case 11:  # HEARTBEAT_ACK
                     logger.debug(f"SHARD ID {self.shard_id}: Heartbeat ACK")