+++ scurrypy/parts/role.py
@@ -7,20 +7,20 @@ from typing import Optional
 class RoleColors(DataModel):
     """Parameters for setting role colors."""
 
-    primary_color: int
+    primary_color: int = None
     """Primary color of the role."""
 
-    secondary_color: int
+    secondary_color: int = None
     """Secondary color of the role. Creates a gradient."""
 
-    tertiary_color: int
+    tertiary_color: int = None
     """Tertiary color of the role. Creates a holographic style."""
 
 @dataclass
 class Role(DataModel):
     """Parameters for creating/editing a role."""
 
-    colors: RoleColors
+    colors: RoleColors = None
     """Colors of the role."""
 
     name: str = None